Encapsulation lesson

+++ chapters/src/exam_revision/CarEntity/CarHealthDiagnosticService.java
@@ -6,7 +6,9 @@ package exam_revision;
  */
 public class CarHealthDiagnosticService {
 
-    public static boolean checkCapacityValue(int capacity) {
+    // D.1.10 Describe how data items can be passed to and from actions as parameters.
+    // I am passing in the capacity so that it can be checked for safety
+    public static boolean checkCapacityValue(float capacity) {
         if (capacity > 0.546) {
             return true;
         }
